Update using_datasets.md
When I run the command '$ floyd run --data GY3QRFFUA8KpbnqvroTPPW "ls -la /input"', floyd will syncing all of data in the current directory. My suggestion is to modify the floyd command not uploading data when the command is just shell command, but not running python script. But now, we can make an empty project to avoid it.