Minor edits to the core concepts page
Switch from third person to second person. This was the only place third person was used, so it was a little off to read.

Also the overview titles seemed superflous, so I cut them.

Let me know if you have already thought about them, just figured it was easier to make a PR to discuss about.